refactor

+++ src/main/java/com/example/springboot/controller/calculadoraApi.java
@@ -3,17 +3,15 @@ package com.example.springboot.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public interface calculadoraApi {
 
     @GetMapping(value = "/calcula")
-    public ResponseEntity<BigInteger> calcular(@RequestParam(name = "operador1") BigInteger primerNumero,
-            @RequestParam(name = "operador2") BigInteger segundoNumero,
+    public ResponseEntity<BigDecimal> calcular(@RequestParam(name = "operador1") BigDecimal primerNumero,
+            @RequestParam(name = "operador2") BigDecimal segundoNumero,
             @RequestParam(name = "operacion") String operacion);
